chore: extract SpecPublisher interface to src/types/publisher.ts
Fixes layering violation where NotionPublisher, NotionSpecCommitter,
OrchestratorImpl, and index.ts imported generic port types from a
Slack-specific adapter file. No behaviour changes.
