+++ src/core/orchestrator.ts
@@ -5,8 +5,8 @@ import { createLogger } from './logger.js';
 import type { SlackAdapter } from '../adapters/slack/slack-adapter.js';
 import type { WorkspaceManager } from './workspace-manager.js';
 import type { SpecGenerator, ReviseResult } from '../adapters/agent/spec-generator.js';
-import { titleFromPath } from '../adapters/slack/canvas-publisher.js';
-import type { SpecPublisher } from '../adapters/slack/canvas-publisher.js';
+import { titleFromPath } from '../types/publisher.js';
+import type { SpecPublisher } from '../types/publisher.js';
 import type { Run, RunStage, RequestIntent } from '../types/runs.js';
 import type { Request, ThreadMessage, InboundEvent } from '../types/events.js';
 import type { FeedbackSource } from '../adapters/notion/notion-feedback-source.js';
